# customer/src/main/java/com/poc/sap/customer/bootstrap/web/CustomerHistoryController.java
package com.poc.sap.customer.bootstrap.web;

import com.poc.sap.common.security.AccessScope;
import com.poc.sap.common.security.ApiRoles;
import com.poc.sap.customer.application.general.CustomerHistoryUseCase;
import com.poc.sap.customer.domain.Customer;
import org.springframework.security.access.prepost.PreAuthorize;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.ExceptionHandler;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;

import java.util.LinkedHashMap;
import java.util.List;
import java.util.Map;
import java.util.NoSuchElementException;

/**
 * Consulta REST del historico de versiones enviadas a SAP (Elasticsearch)
 * y diff entre versiones.
 *
 * <ul>
 *   <li>{@code GET /customers/{id}/history} — versiones (hash + timestamp),
 *       con {@code ?full=true} incluye el snapshot completo de cada una.</li>
 *   <li>{@code GET /customers/{id}/history/diff?from=&to=} — campos que
 *       cambian entre dos versiones; sin parametros compara la ultima contra
 *       la anterior.</li>
 * </ul>
 * Acceso (sdd/common/seguridad-api.md §5): historico con READ o EXTERNAL_READ;
 * a EXTERNAL_READ se le enmascara la PII del snapshot antes de responder; el
 * diff (valores campo a campo) exige READ.
 */
@RestController
@RequestMapping("/customers")
public class CustomerHistoryController {

    private final CustomerHistoryUseCase historyUseCase;
    private final AccessScope accessScope;

    public CustomerHistoryController(CustomerHistoryUseCase historyUseCase, AccessScope accessScope) {
        this.historyUseCase = historyUseCase;
        this.accessScope = accessScope;
    }

    @GetMapping("/{id}/history")
    @PreAuthorize("hasAnyRole('" + ApiRoles.READ + "','" + ApiRoles.EXTERNAL_READ + "')")
    public ResponseEntity<Map<String, Object>> history(
            @PathVariable String id,
            @RequestParam(defaultValue = "false") boolean full) {
        boolean sensitive = accessScope.canSeeSensitiveData();
        List<Map<String, Object>> versions = historyUseCase.history(id).stream()
                .map(s -> {
                    Map<String, Object> v = new LinkedHashMap<>();
                    v.put("payloadHash", s.payloadHash());
                    v.put("timestamp", s.timestamp());
                    if (full) {
                        Customer snapshot = s.entity();
                        v.put("snapshot", sensitive ? snapshot : PiiMasker.mask(snapshot));
                    }
                    return v;
                })
                .toList();
        return ResponseEntity.ok(Map.of(
                "entityId", id,
                "versions", versions,
                "masked", !sensitive));
    }

    @GetMapping("/{id}/history/diff")
    @PreAuthorize("hasRole('" + ApiRoles.READ + "')")
    public ResponseEntity<CustomerHistoryUseCase.HistoryDiff> diff(
            @PathVariable String id,
            @RequestParam(required = false) String from,
            @RequestParam(required = false) String to) {
        return ResponseEntity.ok(historyUseCase.diff(id, from, to));
    }

    @ExceptionHandler(NoSuchElementException.class)
    public ResponseEntity<Map<String, Object>> notFound(NoSuchElementException e) {
        return ResponseEntity.status(HttpStatus.NOT_FOUND)
                .body(Map.of("error", e.getMessage()));
    }
}